Initial Commit
main.py implements the logic of the training. It creates the environment (via which the simulation is controlled) and passes it to the PPO algorithm provided by the stable-baselines library & finally starts training.

customRobotEnv.py contains the implementation of the Gym environment to control the pybullet simulation. It is an adaption of the kukaGymEnv.py, which was copied from the pybullet-library examples. From this file, mainly the main logic defining the Gym env will be kept, but the environment's contents are going to be replaced such that training is performed on a model of a Franka Emika Panda.

kukaGymEnv.py is the original Gym environment from which our own one is going to be derived. It only serves as inspiration, but is not in use. 

Required software can be installed by executing the following pip commands:

pip3 install tensorflow
pip3 install pybullet
pip3 install stable-baselines